+++ errors_data-context.go
@@ -19,14 +19,14 @@ func AddContextDataToError(ctx context.Context, err error) error {
 	return AddDataToError(err, DataFromContext(ctx))
 }
 
-func AddToContext(ctx context.Context, key, value string) context.Context {
+func AddToContext(ctx context.Context, key string, value any) context.Context {
 	v := ctx.Value(dataCtxKey)
 	if v == nil {
-		return context.WithValue(ctx, dataCtxKey, map[string]string{
+		return context.WithValue(ctx, dataCtxKey, map[string]any{
 			key: value,
 		})
 	}
-	data, ok := v.(map[string]string)
+	data, ok := v.(map[string]any)
 	if ok {
 		mutex.Lock()
 		data[key] = value
@@ -35,10 +35,10 @@ func AddToContext(ctx context.Context, key, value string) context.Context {
 	return ctx
 }
 
-func DataFromContext(ctx context.Context) map[string]string {
+func DataFromContext(ctx context.Context) map[string]any {
 	value := ctx.Value(dataCtxKey)
 	if value == nil {
 		return nil
 	}
-	return value.(map[string]string)
+	return value.(map[string]any)
 }